Added code for labeling and filtering applicants
Added code to section 1 to systematically label entries as positive or negative and filter out entries whose CVs are too short to use.

Changed logic for true label: instead of exact match to "Project Manager", an applicant is in the positive class if the position and primary keyword contain "Project Manager".

As of now,
+ positive class = Project Manager
- negative class = QA Engineer
* omitted = All other roles and entries whose CVs are less than 500 characters